Generate (Application)Window builder pattern
This commit adds support for generating the builder pattern for
AdwApplicationWindow and AdwWindow as it makes sense to have them.
It also replaces usage of the ignore_builder object option by
generate_builder = false because ignore_builder was removed from
gir. Since it's expected to have builder patterns for most widgets
or objects, this commit also sets generate_builder = true globally
and special cases the objects where we don't want the builder pattern.